Include block-start ruby annotations when block-fragmenting.
The block-offset of a line box with ruby annotations may have been
increased to make room for the annotations above. Take this into account
when determining whether to break before a line. If the line (including
ruby annotations) is at the block-start of the fragmentainer, we cannot
break before it, as that would prevent content progress, and we'd get
stuck in an infinite loop, retrying to place the line in subsequent
fragmentainers, when the fragmentainers aren't tall enough.

This fixes external/wpt/css/css-ruby/block-ruby-001.html when
LayoutNGBlockFragmentation is enabled. Used to freeze.

Bug: 1253341
Change-Id: I14a5eee3f3b4da6a9bcd5efc664344354755ed55